Inter-task communication is now up and running. There is a static system_data_t variable that can be accessed by all threads. Main thing to worry about is adding critical sections to each thread's attempt to modify the system_data

+++ semi_catkin_ws/src/semi_truck/include/system_data.h
@@ -19,10 +19,10 @@
  */
 struct system_data_t {
    int16_t  wheel_speed;    // speed that the wheel speed sensor is recording
-   int16_t  imu_angle;      // euler angle read by the BNO055 IMU (degrees)
+   float    imu_angle;      // euler angle read by the BNO055 IMU (degrees)
    int16_t  right_URF;      // distance of nearest object for the right URF
-	int16_t  left_URF;       // distance of nearest object for the left URF
-	int16_t  motor_output;   // output sent to the motor driver
+   int16_t  left_URF;       // distance of nearest object for the left URF
+   int16_t  motor_output;   // output sent to the motor driver
    int16_t  steer_output;   // output sent to the steering servo
    int16_t  desired_5th;    // desired state for 5th wheel (locked or unlocked)
    int16_t  actual_5th;     // actual state of the 5th wheel
